pyproject.toml
[project]
name = "texo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.2",
    "datasets==4.0.0",
    "evaluate>=0.4.5",
    "ftfy>=6.3.1",
    "ipykernel>=7.1.0",
    "lightning>=2.5.3",
    "onnxruntime>=1.23.1",
    "opencv-python>=4.10.0.84",
    "optimum-onnx[onnxruntime]>=0.0.1",
    "pillow>=11.1.0",
    "rapidfuzz>=3.13.0",
    "rich>=10.2.2",
    "tensorboard>=2.20.0",
    "timm>=1.0.22",
    "torch==2.7.0",
    "torchvision>=0.20.1",
    "transformers==4.40.0",
    "wand>=0.6.13",
]

[dependency-groups]
dev = [
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "jupyterlab>=4.4.5",
    "matplotlib>=3.10.6",
    "omegaconf>=2.3.0",
    "orjson>=3.11.3",
    "pytest>=8.4.1",
    "texo",
    "tensorboard-plugin-profile>=2.20.6",
    "torch-tb-profiler>=0.4.3",
    "torchinfo>=1.8.0",
    "viztracer>=1.0.4",
    "modelscope>=1.31.0",
]

[tool.uv.pip]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple/"

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
texo = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/texo"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
